Move config calls to somewhere that makes sense
